requirements.txt
sondehub
pyserial
